requirements.txt
Django==1.6.5
Pillow==2.5.1
PyJWT==0.2.1
argparse==1.2.1
conekta==1.1.0
django-carton==1.2.0
django-filebrowser==3.5.6
django-grappelli==2.5.3
django-social-auth==0.7.28
django-taggit==0.12.1
django-tinymce==1.5.2
httplib2==0.9
nose==1.3.4
oauth2==1.5.211
oauthlib==0.6.3
psycopg2==2.5.3
python-openid==2.2.5
python-social-auth==0.2.1
requests==2.4.1
requests-oauthlib==0.4.1
simplejson==3.6.4
six==1.8.0
sorl-thumbnail==11.12
wsgiref==0.1.2
